+++ models/pack.go
@@ -6,7 +6,7 @@ type Pack struct {
 	Base
 	ID              int              `json:"pack_id"`
 	BoxId           int              `json:"box_id"`
-	Date            int              `json:"date"`
+	Date            int64            `json:"date"`
 	Email           string           `json:"email"`
 	TrackingCode    string           `json:"tracking_code"`
 	PackMedications []PackMedication `json:"pack_medications"`
